refactor(cook): move CookMode screens to impl, delegate Child via BlocScreen
CookModeScreen, WhatsCookingScreen, and CookModePreviews move from
cook/public to cook/impl/ui. cook/public retains only the BLoC
interfaces and models; the @Composable Content() body that used to
sit on the CookModeBloc interface as a default is now an override
on CookModeBlocImpl that calls CookModeScreen(this).

RootBloc.Child.CookMode gains `BlocScreen by bloc` so the render
call site is `child.Content()` — Kotlin delegation, no Metro
multi-bindings. Other Child variants are untouched until they
receive the same treatment.

screenshot-test now depends on cook/impl for the preview Blocs;
test bodies use previewCookBlocX.Content() instead of calling
CookModeScreen directly. cook/public sets publicResClass = true so
the moved screens can access the generated Res accessors from impl.
