+++ client/cook/impl/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
     alias(libs.plugins.kotlinSerialization)
 }
 
@@ -10,9 +11,13 @@ kotlin {
             implementation(projects.client.recipe.data.public)
             implementation(projects.client.shared)
             implementation(projects.client.database.core)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(libs.kotlinx.serialization.json)
             implementation(libs.multiplatform.settings)
+            implementation(compose.components.resources)
         }
         commonTest.dependencies { implementation(libs.multiplatform.settings.test) }
     }
