+++ client/cook/public/build.gradle.kts
@@ -20,4 +20,6 @@ kotlin {
     }
 }
 
+compose { resources { publicResClass = true } }
+
 plusLibrary { namespace = "com.plusmobileapps.chefmate.cook" }
