+++ client/cook/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/cook/CookModeBloc.kt
@@ -1,7 +1,5 @@
 package com.plusmobileapps.chefmate.cook
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -25,11 +23,6 @@ interface CookModeBloc : BackClickBloc, BlocScreen {
     fun onLayoutToggled()
 
     fun onKeepScreenOnToggled()
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        CookModeScreen(bloc = this, modifier = modifier)
-    }
 
     enum class LayoutMode {
         Stacked,
