+++ client/ui/screenshot-test/build.gradle.kts
@@ -37,6 +37,7 @@ dependencies {
     implementation(project(":client:auth:ui:public"))
     implementation(project(":client:bottomnav:public"))
     implementation(project(":client:cook:public"))
+    implementation(project(":client:cook:impl"))
     implementation(project(":client:grocery:core:public"))
     implementation(project(":client:recipe:core:public"))
     implementation(project(":client:recipe:list:public"))
